moved hideout function from GameController to HideoutController
