+++ lib/controllers/GameController.js
@@ -7,10 +7,6 @@ const { Quest } = require('../models/Quest');
 const { Locale } = require('../models/Locale');
 const { Trader } = require('../models/Trader');
 const { Item } = require('../models/Item');
-const { HideoutArea } = require('../models/HideoutArea');
-const { HideoutProduction } = require('../models/HideoutProduction');
-const { HideoutScavcase } = require('../models/HideoutScavcase');
-const { Preset } = require('../models/Preset');
 const { RagfairOffer } = require('../models/RagfairOffer');
 const { Ragfair } = require('../models/Ragfair');
 const { getCurrentTimestamp, logger, FastifyResponse, readParsed, generateMongoID } = require("../../utilities");
@@ -464,271 +460,6 @@ class GameController {
         }
     }
 
-    static async clientGameProfileHideoutUpgrade(moveAction = null, _reply = null, playerProfile = null) {
-        logger.logDebug(moveAction);
-        if (playerProfile) {
-            const templateHideoutArea = await HideoutArea.getBy("type", moveAction.areaType);
-            const characterHideoutArea = await playerProfile.character.getHideoutAreaByType(moveAction.areaType);
-
-            if (!templateHideoutArea) {
-                logger.logError(`[clientGameProfileHideoutUpgrade] Upgrading HideoutArea failed. Unknown hideout area ${moveAction.areaType} in hideoutArea database.`);
-                return;
-            }
-
-            if (!characterHideoutArea) {
-                logger.logError(`[clientGameProfileHideoutUpgrade] Upgrading HideoutArea failed. Unknown hideout area ${moveAction.areaType} in character profile.`);
-                return;
-            }
-
-            //logger.logDebug(templateHideoutArea);
-            //logger.logDebug(characterHideoutArea);
-
-            const nextLevel = characterHideoutArea.level + 1;
-            if (typeof templateHideoutArea.stages[nextLevel] === "undefined") {
-                logger.logError(`[clientGameProfileHideoutUpgrade] Upgrading HideoutArea ${templateHideoutArea._id} for character ${playerProfile.character._id} failed. The level ${nextLevel} doesn't exist.`);
-                return;
-            }
-
-            const output = {
-                items: {
-                    new: [],
-                    change: [],
-                    del: []
-                }
-            };
-
-            let allItemsTaken = true;
-            for (const itemToTake of moveAction.items) {
-                const itemTaken = await playerProfile.character.removeItem(itemToTake.id, itemToTake.count);
-                if (itemTaken) {
-                    if (typeof itemTaken.changed !== "undefined") {
-                        output.items.change = output.items.change.concat(itemTaken.changed);
-                    }
-
-                    if (typeof itemTaken.removed !== "undefined") {
-                        output.items.del = output.items.del.concat(itemTaken.removed);
-                    }
-                } else {
-                    allItemsTaken = false;
-                }
-                /*await trader.reduceStock(requestEntry.item_id, requestEntry.count);*/
-            }
-
-            if (allItemsTaken) {
-                const templateHideoutAreaStage = templateHideoutArea.stages[nextLevel];
-                if (templateHideoutAreaStage.constructionTime > 0) {
-                    const currentTime = ~~(Date.now() / 1000);
-                    characterHideoutArea.completeTime = ~~(currentTime + templateHideoutAreaStage.constructionTime);
-                    characterHideoutArea.constructing = true;
-                }
-
-                //logger.logDebug(output);
-                return output;
-            } else {
-                // How do return custom error to client!!1!1!!!111!elf?
-                logger.logError(`[clientGameProfileHideoutUpgrade] Upgrading HideoutArea ${templateHideoutArea._id} for character ${playerProfile.character._id} failed. Unable to take required items.`);
-                return;
-            }
-        }
-    }
-
-    static async clientGameProfileHideoutUpgradeComplete(moveAction = null, _reply = null, playerProfile = null) {
-        if (playerProfile) {
-            const templateHideoutArea = await HideoutArea.getBy("type", moveAction.areaType);
-            const characterHideoutArea = await playerProfile.character.getHideoutAreaByType(moveAction.areaType);
-
-            if (!templateHideoutArea) {
-                logger.logError(`[clientGameProfileHideoutUpgradeComplete] Upgrading HideoutArea failed. Unknown hideout area ${moveAction.areaType} in hideoutArea database.`);
-                return;
-            }
-
-            if (!characterHideoutArea) {
-                logger.logError(`[clientGameProfileHideoutUpgradeComplete] Upgrading HideoutArea failed. Unknown hideout area ${moveAction.areaType} in character profile.`);
-                return;
-            }
-            const nextLevel = characterHideoutArea.level + 1;
-            const templateHideoutAreaStage = templateHideoutArea.stages[nextLevel];
-            if (typeof templateHideoutAreaStage === "undefined") {
-                logger.logError(`[clientGameProfileHideoutUpgradeComplete] Upgrading HideoutArea ${templateHideoutArea._id} for character ${playerProfile.character._id} failed. The level ${nextLevel} doesn't exist.`);
-                return;
-            }
-
-            characterHideoutArea.level = nextLevel;
-            characterHideoutArea.completeTime = 0;
-            characterHideoutArea.constructing = false;
-
-            const hideoutBonuses = templateHideoutAreaStage.bonuses;
-
-            if (typeof hideoutBonuses !== "undefined" && hideoutBonuses.length > 0) {
-                for (const hideoutBonus of hideoutBonuses) {
-                    if (await playerProfile.character.applyHideoutBonus(hideoutBonus)) {
-
-                    }
-                }
-            }
-        }
-    }
-
-    static async clientGameProfileHideoutPutItemsInAreaSlots(moveAction = null, _reply = null, playerProfile = null) {
-        const output = { items: { new: [], change: [], del: [] } };
-        if (playerProfile) {
-            const hideoutArea = await playerProfile.character.getHideoutAreaByType(moveAction.areaType);
-            for (const itemPosition in moveAction.items) {
-                logger.logDebug(moveAction.items);
-                logger.logDebug(itemPosition);
-
-                if (moveAction.items.hasOwnProperty(itemPosition)) {
-                    const itemData = moveAction.items[itemPosition];
-                    const item = await playerProfile.character.getInventoryItemByID(itemData.id);
-                    const slotData = {
-                        item: [
-                            {
-                                _id: item._id,
-                                _tpl: item._tpl,
-                                upd: item.upd
-                            }
-                        ]
-                    };
-                    hideoutArea.slots[itemPosition] = slotData;
-                    await playerProfile.character.removeItem(item._id);
-                    output.items.del.push(item);
-                }
-            }
-        }
-        return output;
-    }
-
-    static async clientGameProfileHideoutTakeItemsFromAreaSlots(moveAction = null, _reply = null, playerProfile = null) {
-        const output = { items: { new: [], change: [], del: [] } };
-        if (playerProfile) {
-            const hideoutArea = await playerProfile.character.getHideoutAreaByType(moveAction.areaType);
-            if (!hideoutArea) {
-                logger.logError(`[clientGameProfileHideoutTakeItemsFromAreaSlots] Unable to find hideout area type ${moveAction.areaType} for playerProfile ${playerProfile.character._id}.`);
-                return output;
-            }
-
-            for (const slot in moveAction.slots) {
-                for (const item of hideoutArea.slots[slot].item) {
-                    const itemAdded = await playerProfile.character.addItem(await playerProfile.character.getStashContainer(), item._tpl, false, 1);
-                    if (itemAdded) {
-                        output.items.new = [...output.items.new, ...itemAdded];
-                        hideoutArea.slots.splice(slot, 1);
-                    }
-                }
-            }
-        }
-        return output;
-    }
-
-    static async clientGameProfileHideoutToggleArea(moveAction = null, _reply = null, playerProfile = null) {
-        if (playerProfile) {
-            const hideoutArea = await playerProfile.character.getHideoutAreaByType(moveAction.areaType);
-            if (!hideoutArea) {
-                logger.logError(`[clientGameProfileHideoutToggleArea] Unable to find hideout area type ${moveAction.areaType} for playerProfile ${playerProfile.character._id}.`);
-                return;
-            }
-            hideoutArea.active = moveAction.enabled;
-        }
-    }
-
-    static async clientGameProfileHideoutSingleProductionStart(moveAction = null, _reply = null, playerProfile = null) {
-        logger.logDebug(moveAction);
-        if (playerProfile) {
-            const hideoutProductionTemplate = await HideoutProduction.get(moveAction.recipeId);
-            if (!hideoutProductionTemplate) {
-                logger.logError(`[clientGameProfileHideoutSingleProductionStart] Starting hideout production failed. Unknown hideout production with Id ${moveAction.recipeId} in hideoutProduction database.`);
-                return;
-            }
-
-            const output = {
-                items: {
-                    new: [],
-                    change: [],
-                    del: []
-                }
-            };
-
-            let allItemsTaken = true;
-            for (const itemToTake of moveAction.items) {
-                const itemTaken = await playerProfile.character.removeItem(itemToTake.id, itemToTake.count);
-                if (itemTaken) {
-                    if (typeof itemTaken.changed !== "undefined") {
-                        output.items.change = output.items.change.concat(itemTaken.changed);
-                    }
-
-                    if (typeof itemTaken.removed !== "undefined") {
-                        output.items.del = output.items.del.concat(itemTaken.removed);
-                    }
-                } else {
-                    allItemsTaken = false;
-                }
-                /*await trader.reduceStock(requestEntry.item_id, requestEntry.count);*/
-            }
-
-            if (allItemsTaken) {
-                let productionTime = 0;
-
-                if (typeof hideoutProductionTemplate.ProductionTime !== "undefined") {
-                    productionTime = hideoutProductionTemplate.ProductionTime;
-                } else if (typeof hideoutProductionTemplate.productionTime !== "undefined") {
-                    productionTime = hideoutProductionTemplate.productionTime;
-                }
-
-                if (!hideoutProductionTemplate.count) {
-                    hideoutProductionTemplate.count = 1;
-                }
-
-                const products = [{
-                    _id: await generateMongoID(),
-                    _tpl: hideoutProductionTemplate.endProduct,
-                    count: hideoutProductionTemplate.count
-                }];
-
-                playerProfile.character.Hideout.Production[hideoutProductionTemplate._id] = {
-                    Progress: 0,
-                    inProgress: true,
-                    Products: products,
-                    RecipeId: moveAction.recepieId,
-                    SkipTime: 0,
-                    ProductionTime: parseInt(productionTime),
-                    StartTimestamp: await getCurrentTimestamp()
-                };
-
-                return output;
-            } else {
-                // How do return custom error to client!!1!1!!!111!elf?
-                logger.logError(`[clientGameProfileHideoutSingleProductionStart] Starting hideout production for recepie with Id ${moveAction.recipeId} failed. Unable to take required items.`);
-                return;
-            }
-        }
-    }
-
-    static async clientGameProfileHideoutContinuousProductionStart(moveAction = null, _reply = null, playerProfile = null) {
-        if (playerProfile) {
-            const hideoutProductionTemplate = await HideoutProduction.get(moveAction.recipeId);
-            if (!hideoutProductionTemplate) {
-                logger.logError(`[clientGameProfileHideoutContinuousProductionStart] Couldn't start hideout production. Unknown production with Id ${moveAction.recipeId}`);
-                return;
-            }
-
-            let productionTime = 0
-            if (typeof hideoutProductionTemplate.ProductionTime !== "undefined") {
-                productionTime = hideoutProductionTemplate.ProductionTime;
-            } else if (typeof hideoutProductionTemplate.productionTime !== "undefined") {
-                productionTime = hideoutProductionTemplate.productionTime;
-            }
-
-            playerProfile.character.Hideout.Production[hideoutProductionTemplate._id] = {
-                Progress: 0,
-                inProgress: true,
-                RecipeId: moveAction.recipeId,
-                SkipTime: 0,
-                ProductionTime: parseInt(productionTime),
-                StartTimestamp: await getCurrentTimestamp()
-            };
-        }
-    };
-
     static async clientGameProfileAddNote(moveAction = null, _reply = null, playerProfile = null) {
         if (playerProfile) {
             const playerPMC = await playerProfile.getPmc();
@@ -771,81 +502,6 @@ class GameController {
             playerPMC.WishList = [];
             await playerPMC.save();
         }
-    }
-
-    static async clientGameProfileHideoutScavCaseProductionStart(moveAction = null, _reply = null, playerProfile = null) {
-        const output = {
-            items: {
-                new: [],
-                change: [],
-                del: []
-            }
-        };
-        if (playerProfile) {
-            const hideoutScavcaseProduction = await HideoutScavcase.get(moveAction.recipeId);
-            if (!hideoutScavcaseProduction) {
-                logger.logError(`[clientGameProfileHideoutScavCaseProductionStart] Couldn't start scavcase. Unknown hideout scavcase with Id ${moveAction.recipeId}`);
-            }
-            const itemTaken = await playerProfile.character.removeItem(moveAction.items[0].id, moveAction.items[0].count);
-
-            const products = await hideoutScavcaseProduction.generateRewards();
-
-            if (itemTaken) {
-                output.items.change = itemTaken.changed;
-                output.items.removed = itemTaken.removed;
-                playerProfile.character.Hideout.Production[hideoutScavcaseProduction._id] = {
-                    Progress: 0,
-                    inProgress: true,
-                    RecipeId: moveAction.recipeId,
-                    Products: products,
-                    SkipTime: 0,
-                    ProductionTime: parseInt(hideoutScavcaseProduction.ProductionTime),
-                    StartTimestamp: await getCurrentTimestamp()
-                };
-            } else {
-                logger.logError(`[clientGameProfileHideoutScavCaseProductionStart] Couldn't take money with id ${moveAction.items[0].id}`);
-            }
-        }
-        return output;
-    }
-
-    static async clientGameProfileHideoutTakeProduction(moveAction = null, _reply = null, playerProfile = null) {
-        const output = {
-            items: {
-                new: [],
-                change: [],
-                del: []
-            }
-        };
-        // TODO: HANDLE STACK FOR BULLETS & BULLETS PACKS
-        if (playerProfile) {
-            let itemsAdded;
-            const production = await playerProfile.character.getHideoutProductionById(moveAction.recipeId);
-            if (!production.hasOwnProperty("Products")) {
-                logger.logError(`[clientGameProfileHideoutTakeProduction] Remanent productions error: no products for production with Id ${moveAction.recipeId}`);
-                await playerProfile.character.removeHideoutProductionById(moveAction.recipeId);
-                return output;
-            }
-            for (const product of production.Products) {
-                if (!product.count) {
-                    product.count = 1;
-                }
-                const itemTemplate = await Item.get(product._tpl);
-                if (await Preset.itemHasPreset(itemTemplate._id)) {
-                    const itemPresets = await Preset.getPresetsForItem(itemTemplate._id);
-                    const itemPreset = Object.values(itemPresets).find(preset => preset._encyclopedia);
-                    const basedChildren = await Item.prepareChildrenForAddItem(itemPreset._items[0], itemPreset._items);
-                    itemsAdded = await playerProfile.character.addItem(await playerProfile.character.getStashContainer(), itemTemplate._id, basedChildren, product.count, true);
-                } else {
-                    itemsAdded = await playerProfile.character.addItem(await playerProfile.character.getStashContainer(), itemTemplate._id, undefined, product.count, true);
-                }
-                if (itemsAdded) {
-                    output.items.new = output.items.new.concat(itemsAdded);
-                }
-            }
-            await playerProfile.character.removeHideoutProductionById(moveAction.recipeId);
-        }
-        return output;
     }
 
     static async clientGameProfileCustomizationBuy(moveAction = null, _reply = null, playerProfile = null) {
