+++ lib/controllers/index.js
@@ -9,6 +9,7 @@ const AccountController = require("./AccountController");
 const WeblauncherController = require("./WeblauncherController");
 const RaidController = require("./RaidController");
 const ItemController = require("./ItemController");
+const HideoutController = require("./HideoutController");
 
 module.exports = {
     ...AccountController,
@@ -21,5 +22,6 @@ module.exports = {
     ...TradingController,
     ...WeblauncherController,
     ...RaidController,
-    ...ItemController
+    ...ItemController,
+    ...HideoutController
 };
